requirements.txt
pytorch-lightning
pandas
scikit-learn
h5py
pytorch3d